Implements mergerepo_c support for --koji simple mode
